Added the Schedule Display
Currently have a label at top "Weekly Schedule" which might not be necessary. I was able to add pagination but currently only have one image as schedule. I attempted to customize the pagination to match the theme but I found it very limiting. However, I managed to place the pagination at the top of the content. Lastly, created a basic color "legend" for the course type/method at the bottom of the schedule.